First fix-up pass.
Untabified the SessionFinder files and fixed the indentation errors.

There was a compile error related to the static ints at the start of the header.
I went ahead and changed them to #defines instead of static const ints because
they're going to get shoved into the .bss data anyway, and this way saves a
compile-time constant folding pass.  It's also the standard way of defining
things.

Also fixed a few misc compile errors.  It still doesn't compile because the code
is assuming that we've done things that we haven't, but it's better at least.