Fix tracker request and response handling.
Previously we were not handling regular tracker requests, which are used to
retrieve more peers. This breaks torrents that run for long periods of time.
Previously long-lived torrents were not adding all the pieces even though they
were all in the capture because the peer they were recieving the data from was
not recognized.